refactor: refactor item to new file
