+++ TypeScript/app/gilded-rose.ts
@@ -1,14 +1,4 @@
-export class Item {
-  name: string;
-  sellIn: number;
-  quality: number;
-
-  constructor(name, sellIn, quality) {
-    this.name = name;
-    this.sellIn = sellIn;
-    this.quality = quality;
-  }
-}
+import { Item } from "./item";
 
 export class GildedRose {
   items: Array<Item>;
