+++ TypeScript/test/golden-master-text-test.ts
@@ -1,4 +1,5 @@
-import { Item, GildedRose } from '../app/gilded-rose';
+import { Item } from '../app/item';
+import { GildedRose } from '../app/gilded-rose';
 
 const items = [
   new Item("+5 Dexterity Vest", 10, 20), //
@@ -17,8 +18,8 @@ const gildedRose = new GildedRose(items);
 
 let days: number = 2;
 if (process.argv.length > 2) {
-    days = +process.argv[2];
-  }
+  days = +process.argv[2];
+}
 
 for (let i = 0; i < days; i++) {
   console.log("-------- day " + i + " --------");
