+++ TypeScript/test/jest/gilded-rose.spec.ts
@@ -1,4 +1,5 @@
-import { Item, GildedRose } from '@/gilded-rose';
+import { GildedRose } from '@/gilded-rose';
+import { Item } from '@/item';
 
 describe('Gilded Rose', () => {
   it('should foo', () => {
