+++ TypeScript/test/mocha/gilded-rose.spec.ts
@@ -1,5 +1,6 @@
 import { expect } from 'chai';
-import { Item, GildedRose } from '@/gilded-rose';
+import { GildedRose } from '@/gilded-rose';
+import { Item } from '@/item';
 
 describe('Gilded Rose', () => {
   it('should foo', () => {
